+++ crates/adapters/mock-da/src/storable/rpc/server.rs
@@ -232,65 +232,3 @@ pub async fn start_server(da_service: StorableMockDaService) -> SocketAddr {
     tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
     addr
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{BlockProducingConfig, MockAddress};
-    use axum::http::StatusCode;
-    use axum_test::TestServer;
-    use std::sync::Arc;
-    use tokio::sync::RwLock;
-
-    use crate::storable::layer::StorableMockDaLayer;
-
-    #[tokio::test]
-    async fn test_get_head_block_header() {
-        let da_layer = Arc::new(RwLock::new(
-            StorableMockDaLayer::new_in_memory(0)
-                .await
-                .expect("Failed to create DA layer"),
-        ));
-        let da_service = StorableMockDaService::new(
-            MockAddress::new([1; 32]),
-            da_layer,
-            //BlockProducingConfig::Periodic { block_time_ms: 100 },
-            BlockProducingConfig::Manual,
-        )
-        .await;
-
-        da_service.produce_n_blocks_now(10).await;
-
-        let app = create_router(da_service);
-        let server = TestServer::new(app).unwrap();
-
-        let response = server.get("/head-block-header").await;
-        println!("{:?}, ", response);
-        assert_eq!(response.status_code(), StatusCode::OK);
-    }
-
-    #[tokio::test]
-    async fn test_send_transaction() {
-        let da_layer = Arc::new(RwLock::new(
-            StorableMockDaLayer::new_in_memory(0)
-                .await
-                .expect("Failed to create DA layer"),
-        ));
-        let da_service = StorableMockDaService::new(
-            MockAddress::new([1; 32]),
-            da_layer,
-            BlockProducingConfig::Periodic { block_time_ms: 100 },
-        )
-        .await;
-
-        let app = create_router(da_service);
-        let server = TestServer::new(app).unwrap();
-
-        let request = SubmitTransactionRequest {
-            blob: hex::encode(b"test blob"),
-        };
-
-        let response = server.post("/send-transaction").json(&request).await;
-        assert_eq!(response.status_code(), StatusCode::OK);
-    }
-}
